refactor: Image(main, description) 통합 및 코드 리팩토링

+++ src/main/java/com/codenear/butterfly/admin/products/application/AdminProductService.java
@@ -52,16 +52,14 @@ public class AdminProductService {
 
         ProductInventory product = ProductInventory.builder()
                 .createRequest(request)
-                .productImage(imageConverter(request.productImage()))
                 .deliveryInformation(deliveryInformation)
                 .keywords(keywords)
                 .build();
         productRepository.save(product);
         kakaoPaymentRedisRepository.saveStockQuantity(request.productName(), request.stockQuantity());
-        if (request.descriptionImages().get(0) != null && !request.descriptionImages().get(0).isEmpty()) {
-            List<ProductImage> descriptionImages = getDescriptionImages(request.descriptionImages(), product);
-            productImageRepository.saveAll(descriptionImages);
-        }
+
+        saveImage(request.productImage(), product, ProductImage.ImageType.MAIN);
+        saveImage(request.descriptionImages(), product, ProductImage.ImageType.DESCRIPTION);
     }
 
     public List<ProductInventory> loadAllProducts() {
@@ -72,17 +70,8 @@ public class AdminProductService {
     public void updateProduct(Long id, ProductUpdateRequest request) {
         Product product = findById(id);
 
-        if (request.getProductImage() != null && !request.getProductImage().isEmpty()) {
-            if (product.getProductImage() != null) {
-                deleteS3UploadedFile(product);
-            }
-            product.setProductImage(imageConverter(request.getProductImage()));
-        }
-
-        // 상품 설명 이미지 업데이트
-        if (request.getDescriptionImages().get(0) != null && !request.getDescriptionImages().get(0).isEmpty()) {
-            updateDescriptionImages(request, product);
-        }
+        updateImage(request.getProductImage(), product, ProductImage.ImageType.MAIN);
+        updateImage(request.getDescriptionImages(), product, ProductImage.ImageType.DESCRIPTION);
 
         product.update(request);
         kakaoPaymentRedisRepository.saveStockQuantity(product.getProductName(), request.getStockQuantity());
@@ -145,30 +134,38 @@ public class AdminProductService {
         return null;
     }
 
-    private void deleteS3UploadedFile(Product product) {
-        String existingFileName = extractFileNameFromUrl(product.getProductImage());
+    private void deleteS3UploadedFile(String imageUrl) {
+        String existingFileName = extractFileNameFromUrl(imageUrl);
         s3Service.deleteFile(existingFileName, PRODUCT_IMAGE);
     }
 
-    private List<ProductImage> getDescriptionImages(List<MultipartFile> images, Product product) {
-        return images.stream()
-                .map(image -> ProductImage.builder()
-                        .imageType(ProductImage.ImageType.DESCRIPTION)
-                        .imageUrl(imageConverter(image))
-                        .product(product)
-                        .build())
-                .toList();
+    private List<ProductImage> saveImage(List<MultipartFile> images, Product product, ProductImage.ImageType imageType) {
+        if (!images.get(0).isEmpty() && images.get(0) != null) {
+            List<ProductImage> productImages = images.stream()
+                    .map(image -> ProductImage.builder()
+                            .imageType(imageType)
+                            .imageUrl(imageConverter(image))
+                            .product(product)
+                            .build())
+                    .toList();
+            return productImageRepository.saveAll(productImages);
+        }
+        return null;
     }
 
-    private void updateDescriptionImages(ProductUpdateRequest request, Product product) {
-        List<ProductImage> descriptionImages = productImageRepository.findAllByProductIdAndImageType(product.getId(), ProductImage.ImageType.DESCRIPTION);
-        descriptionImages
-                .forEach(image -> {
-                    deleteS3UploadedFile(product);
-                    productImageRepository.deleteById(image.getId());
-                });
-        List<ProductImage> newDescriptionImages = getDescriptionImages(request.getDescriptionImages(), product);
-        productImageRepository.saveAll(newDescriptionImages);
-        product.updateDescriptionImage(newDescriptionImages);
+    private void updateImage(List<MultipartFile> images, Product product, ProductImage.ImageType imageType) {
+        if (!images.get(0).isEmpty() && images.get(0) != null) {
+            List<ProductImage> descriptionImages = productImageRepository.findAllByProductIdAndImageType(product.getId(), imageType);
+            descriptionImages
+                    .forEach(image -> {
+                        deleteS3UploadedFile(image.getImageUrl());
+                        productImageRepository.deleteById(image.getId());
+                    });
+            List<ProductImage> newImages = saveImage(images, product, imageType);
+            switch (imageType) {
+                case MAIN -> product.updateMainImage(newImages);
+                case DESCRIPTION -> product.updateDescriptionImage(newImages);
+            }
+        }
     }
 }
